Add comment and move pillars.ts
